requirements.txt
_libgcc_mutex==0.1
astroid==2.4.2
blas==1.0
ca-certificates==2020.12.5
certifi==2020.12.5
cudatoolkit==10.2.89
cycler==0.10.0
freetype==2.10.3
icu==67.1
intel-openmp==2020.2
isort==5.7.0
jpeg==9b
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lcms2==2.11
ld_impl_linux-64==2.33.1
libedit==3.1.20191231
libffi==3.3
libgcc-ng==9.1.0
libgfortran-ng==7.5.0
libgfortran4==7.5.0
libpng==1.6.37
libstdcxx-ng==9.1.0
libtiff==4.1.0
lz4-c==1.9.2
matplotlib==3.2.2
matplotlib-base==3.2.2
mccabe==0.6.1
mkl==2020.2
mkl-service==2.3.0
mkl_fft==1.2.0
mkl_random==1.1.1
ncurses==6.2
ninja==1.10.1
numpy==1.19.1
numpy-base==1.19.1
olefile==0.46
openssl==1.1.1j
pandas==1.1.3
pillow==7.2.0
pip==20.2.3
pylint==2.6.0
pyparsing==2.4.7
python==3.8.5
python-dateutil==2.8.1
python_abi==3.8
pytorch==1.6.0
pytz==2020.1
readline==8.0
scipy==1.6.0
setuptools==50.3.0
six==1.15.0
sqlite==3.33.0
tk==8.6.10
toml==0.10.1
torchvision==0.7.0
tornado==6.1
wheel==0.35.1
wrapt==1.11.2
xz==5.2.5
zlib==1.2.11
zstd==1.4.5
opencv-python==4.5.1.48
